third_party/WebKit/LayoutTests/http/tests/credentialmanager/localcredential-basics.html

<!DOCTYPE html>
<title>Credential Manager: LocalCredential basics.</title>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script src="/serviceworker/resources/interfaces.js"></script>
<script>
test(function() {
    var credential = new LocalCredential('id', 'name', 'https://example.com/avatar.png', 'pencil');

    verifyInterface('LocalCredential', credential, {
        id: 'string',
        name: 'string',
        avatarURL: 'string',
        password: 'string'
    });

    assert_equals(credential.id, 'id');
    assert_equals(credential.name, 'name');
    assert_equals(credential.avatarURL, 'https://example.com/avatar.png');
    assert_equals(credential.password, 'pencil');
}, 'Interfaces and attributes of LocalCredential');

test(function() {
    assert_throws(new SyntaxError(), function () {
        var credential = new LocalCredential('id', 'name', '-', 'pencil');
    });
}, 'Construct a LocalCredential with an invalid avatar URL.');
</script>
